requirements.txt
django-widget-tweaks==1.4.12
pillow
setuptools
requests